refactor: modularize MedicineForm component and improve validation
